refactor: refactor hot keys handler
