+++ src/store/selectors.js
@@ -9,12 +9,10 @@ const gameOnSelector = (state) => state.gameReducer.isGameOn;
 const gameMovementSelector = (state) => state.gameReducer.isMoving;
 
 const authSelector = (state) => state.firebaseReducer.auth;
-const fireStoreLevelSelector = (state) => state.firestoreReducer.level;
 const userNameSelector = (state) => state.firebaseReducer.profile.displayName;
 
 export {
   authSelector,
-  fireStoreLevelSelector,
   userNameSelector,
   musicVolumeSelector,
   soundVolumeSelector,
